컨트롤러 수정 (예약 제외)

+++ wheere/src/main/java/com/d138/wheere/controller/memberDTO/RateDriverDTO.java
@@ -4,5 +4,5 @@ import lombok.Data;
 @Data
 public class RateDriverDTO {
     private Long rId;
-    private Long rate;
+    private Double rate;
 }
